LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.nonfinite-expected.txt

2d.drawImage.nonfinite

drawImage() with Infinity/NaN is ignored


PASS drawImage() with Infinity/NaN is ignored 
